CA-325582: remove unused Host datasources to avoid space leaks
This needs to be deployed with corresponding commit in rrdd-plugins.

Previously we always kept all datasources we've ever seen for a host
even if we temporarily stopped receiving RRD updates for it.
This is useful for per-host SR-level RRDs which only exist when at least
one VM using it is running: we want to see the historic data when the VM
is started again.
However if SRs are attached/detached frequently this can cause xcp-rrdd
to increase in memory usage over time, even across toolstack restarts or
host reboots, because all host-level RRDs were preserved.
This was not visible in `xe host-data-source-list`, since that lists
only the active RRDs.

Solution is to preserve only the latest reported RRDs for hosts, and to
ask plugins to keep reporting data for inactive RRDs that they need.
E.g. rrdp-iostat should keep reporting data for SRs that have a PBD for
that host, even if no VM is using it or the PBD is temporarily
unplugged.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>